Added backendTCP example test
