+++ __tests__/backendTCP/local/busses/bus.0/bus.yaml
@@ -1,0 +1,3 @@
+host: localhost
+port: 3007
+timeout: 200
