+++ __tests__/backendTCP/local/busses/bus.0/s1.yaml
@@ -1,0 +1,2 @@
+slaveid: 1
+specificationid: dimplexpco5
