+++ __tests__/backendTCP/local/secrets.yaml
@@ -1,0 +1,5 @@
+mqttpassword: modbus2mqtt
+mqttuser: modbus2mqtt
+githubPersonalToken: someToken
+username: test
+password: $2a$08$9HpW0t2gJ6KddJWgDmdH1OyaYKucIAZpKjUFZArkaAEkTWcH7ORMq
