+++ __tests__/backendTCP/local/specifications/files/dimplexpco5/files.yaml
@@ -1,0 +1,6 @@
+- url: https://dimplex.atlassian.net/wiki/spaces/DW/pages/2862481429/Modbus+RTU+Anbindung%20pdf
+  fileLocation: 1
+  usage: doc
+- url: /specifications/files/dimplexpco5/dimplexpco5.jpg
+  fileLocation: 0
+  usage: img
